changed from float to double, working with T_rec spacepoints now

+++ ubreco/WcpPortedReco/ProducePort/SimpleSpacePoint.h
@@ -1,6 +1,6 @@
 struct SimpleSpacePoint {
-  float x;
-  float y;
-  float z;
-  float q;
+  double x;
+  double y;
+  double z;
+  double q;
 };
